Split typechecker core test modules
