+++ src/typechecker/tests.rs
@@ -14,1251 +14,11 @@ mod resolver_metadata;
 
 mod resolver_validation;
 
-#[test]
-fn value_signature_absence_validation_builds_entries() {
-    let program = parse_program(
-        r#"
-add = (left: i32, right: i32) i32 { left + right }
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let symbol = symbols
-        .lookup(Namespace::Value, "add")
-        .expect("value symbol");
-    let entries = ValueSignatureAbsenceValidation {
-        parameter_count_code: "PARAM_COUNT",
-        parameter_name_code: "PARAM_NAMES",
-        parameter_type_name_code: "PARAM_TYPES",
-        parameter_type_code: "TYPED_PARAM_TYPES",
-        return_type_code: "RETURN_TYPE",
-        typed_return_type_code: "TYPED_RETURN_TYPE",
-    }
-    .entries(symbol);
-
-    assert!(entries.iter().all(|entry| entry.present));
-    assert_eq!(
-        entries.map(|entry| entry.message("value", "add")),
-        [
-            "resolver value symbol 'add' has parameter count metadata, expected none",
-            "resolver value symbol 'add' has parameter names metadata, expected none",
-            "resolver value symbol 'add' has parameter types metadata, expected none",
-            "resolver value symbol 'add' has typed parameter types metadata, expected none",
-            "resolver value symbol 'add' has return type metadata, expected none",
-            "resolver value symbol 'add' has typed return type metadata, expected none",
-        ]
-    );
-}
-
-#[test]
-fn value_signature_absence_validation_uses_module_resolver_codes() {
-    let validation = ValueSignatureAbsenceValidation::module_resolver_codes();
-
-    assert_eq!(validation.parameter_count_code, "E0265");
-    assert_eq!(validation.parameter_name_code, "E0267");
-    assert_eq!(validation.parameter_type_name_code, "E0268");
-    assert_eq!(validation.parameter_type_code, "E0371");
-    assert_eq!(validation.return_type_code, "E0266");
-    assert_eq!(validation.typed_return_type_code, "E0372");
-}
-
-#[test]
-fn value_signature_absence_validation_uses_import_resolver_codes() {
-    let validation = ValueSignatureAbsenceValidation::import_resolver_codes();
-
-    assert_eq!(validation.parameter_count_code, "E0281");
-    assert_eq!(validation.parameter_name_code, "E0283");
-    assert_eq!(validation.parameter_type_name_code, "E0284");
-    assert_eq!(validation.parameter_type_code, "E0362");
-    assert_eq!(validation.return_type_code, "E0282");
-    assert_eq!(validation.typed_return_type_code, "E0363");
-}
-
-#[test]
-fn value_signature_absence_validation_uses_local_resolver_codes() {
-    let validation = ValueSignatureAbsenceValidation::local_resolver_codes();
-
-    assert_eq!(validation.parameter_count_code, "E0249");
-    assert_eq!(validation.parameter_name_code, "E0251");
-    assert_eq!(validation.parameter_type_name_code, "E0252");
-    assert_eq!(validation.parameter_type_code, "E0380");
-    assert_eq!(validation.return_type_code, "E0250");
-    assert_eq!(validation.typed_return_type_code, "E0381");
-}
-
-#[test]
-fn value_signature_absence_validation_uses_type_like_resolver_codes() {
-    let validation = ValueSignatureAbsenceValidation::type_like_resolver_codes();
-
-    assert_eq!(validation.parameter_count_code, "E0310");
-    assert_eq!(validation.parameter_name_code, "E0312");
-    assert_eq!(validation.parameter_type_name_code, "E0313");
-    assert_eq!(validation.parameter_type_code, "E0360");
-    assert_eq!(validation.return_type_code, "E0311");
-    assert_eq!(validation.typed_return_type_code, "E0361");
-}
-
-#[test]
-fn value_signature_absence_validation_uses_variant_resolver_codes() {
-    let validation = ValueSignatureAbsenceValidation::variant_resolver_codes();
-
-    assert_eq!(validation.parameter_count_code, "E0330");
-    assert_eq!(validation.parameter_name_code, "E0332");
-    assert_eq!(validation.parameter_type_name_code, "E0333");
-    assert_eq!(validation.parameter_type_code, "E0389");
-    assert_eq!(validation.return_type_code, "E0331");
-    assert_eq!(validation.typed_return_type_code, "E0390");
-}
-
-#[test]
-fn type_parameter_absence_validation_builds_entries() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    encode: (Self) str
-}
-
-identity<T: Json> = (value: T) T { value }
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let symbol = symbols
-        .lookup(Namespace::Value, "identity")
-        .expect("value symbol");
-    let entries = TypeParameterAbsenceValidation {
-        count_code: "COUNT",
-        name_code: "NAMES",
-        bound_code: "BOUNDS",
-        bound_ref_code: "BOUND_REFS",
-    }
-    .entries(symbol);
-
-    assert_eq!(
-        entries,
-        [
-            AbsentMetadataEntry::new(true, "COUNT", "type parameter count"),
-            AbsentMetadataEntry::new(true, "NAMES", "type parameter names"),
-            AbsentMetadataEntry::new(true, "BOUNDS", "type parameter bounds"),
-            AbsentMetadataEntry::new(true, "BOUND_REFS", "typed type parameter bound refs"),
-        ]
-    );
-}
-
-#[test]
-fn type_parameter_absence_validation_uses_module_resolver_codes() {
-    let validation = TypeParameterAbsenceValidation::module_resolver_codes();
-
-    assert_eq!(validation.count_code, "E0269");
-    assert_eq!(validation.name_code, "E0348");
-    assert_eq!(validation.bound_code, "E0270");
-    assert_eq!(validation.bound_ref_code, "E0373");
-}
-
-#[test]
-fn type_parameter_absence_validation_uses_import_resolver_codes() {
-    let validation = TypeParameterAbsenceValidation::import_resolver_codes();
-
-    assert_eq!(validation.count_code, "E0285");
-    assert_eq!(validation.name_code, "E0349");
-    assert_eq!(validation.bound_code, "E0286");
-    assert_eq!(validation.bound_ref_code, "E0364");
-}
-
-#[test]
-fn type_parameter_absence_validation_uses_local_resolver_codes() {
-    let validation = TypeParameterAbsenceValidation::local_resolver_codes();
-
-    assert_eq!(validation.count_code, "E0253");
-    assert_eq!(validation.name_code, "E0350");
-    assert_eq!(validation.bound_code, "E0254");
-    assert_eq!(validation.bound_ref_code, "E0382");
-}
-
-#[test]
-fn type_parameter_absence_validation_uses_variant_resolver_codes() {
-    let validation = TypeParameterAbsenceValidation::variant_resolver_codes();
-
-    assert_eq!(validation.count_code, "E0334");
-    assert_eq!(validation.name_code, "E0351");
-    assert_eq!(validation.bound_code, "E0335");
-    assert_eq!(validation.bound_ref_code, "E0391");
-}
-
-#[test]
-fn field_absence_validation_builds_entries() {
-    let program = parse_program(
-        r#"
-Point: { x: i32, y: i32 }
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let symbol = symbols
-        .lookup(Namespace::Type, "Point")
-        .expect("type symbol");
-    let entries = FieldAbsenceValidation {
-        count_code: "COUNT",
-        type_name_code: "FIELD_TYPES",
-        typed_code: "TYPED_FIELDS",
-    }
-    .entries(symbol);
-
-    assert_eq!(
-        entries,
-        [
-            AbsentMetadataEntry::new(true, "COUNT", "field count"),
-            AbsentMetadataEntry::new(true, "FIELD_TYPES", "field types"),
-            AbsentMetadataEntry::new(true, "TYPED_FIELDS", "typed field types"),
-        ]
-    );
-}
-
-#[test]
-fn field_absence_validation_uses_module_resolver_codes() {
-    let validation = FieldAbsenceValidation::module_resolver_codes();
-
-    assert_eq!(validation.count_code, "E0271");
-    assert_eq!(validation.type_name_code, "E0272");
-    assert_eq!(validation.typed_code, "E0374");
-}
-
-#[test]
-fn field_absence_validation_uses_import_resolver_codes() {
-    let validation = FieldAbsenceValidation::import_resolver_codes();
-
-    assert_eq!(validation.count_code, "E0287");
-    assert_eq!(validation.type_name_code, "E0288");
-    assert_eq!(validation.typed_code, "E0365");
-}
-
-#[test]
-fn field_absence_validation_uses_local_resolver_codes() {
-    let validation = FieldAbsenceValidation::local_resolver_codes();
-
-    assert_eq!(validation.count_code, "E0255");
-    assert_eq!(validation.type_name_code, "E0256");
-    assert_eq!(validation.typed_code, "E0383");
-}
-
-#[test]
-fn field_absence_validation_uses_type_like_resolver_codes() {
-    let validation = FieldAbsenceValidation::type_like_resolver_codes();
-
-    assert_eq!(validation.count_code, "E0319");
-    assert_eq!(validation.type_name_code, "E0320");
-    assert_eq!(validation.typed_code, "E0398");
-}
-
-#[test]
-fn field_absence_validation_uses_variant_resolver_codes() {
-    let validation = FieldAbsenceValidation::variant_resolver_codes();
-
-    assert_eq!(validation.count_code, "E0336");
-    assert_eq!(validation.type_name_code, "E0337");
-    assert_eq!(validation.typed_code, "E0392");
-}
-
-#[test]
-fn field_absence_validation_uses_behavior_resolver_codes() {
-    let validation = FieldAbsenceValidation::behavior_resolver_codes();
-
-    assert_eq!(validation.count_code, "E0321");
-    assert_eq!(validation.type_name_code, "E0322");
-    assert_eq!(validation.typed_code, "E0399");
-}
-
-#[test]
-fn field_absence_validation_uses_value_resolver_codes() {
-    let validation = FieldAbsenceValidation::value_resolver_codes();
-
-    assert_eq!(validation.count_code, "E0298");
-    assert_eq!(validation.type_name_code, "E0299");
-    assert_eq!(validation.typed_code, "E0403");
-}
-
-#[test]
-fn variant_absence_validation_builds_entries() {
-    let program = parse_program(
-        r#"
-Option<T>: Some(T), None
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let symbol = symbols
-        .lookup_variant("Option", "Some")
-        .expect("variant symbol");
-    let entries = VariantAbsenceValidation {
-        names_code: "NAMES",
-        owner_code: "OWNER",
-        payload_count_code: "PAYLOAD_COUNT",
-        payload_type_name_code: "PAYLOAD_TYPE",
-        payload_type_code: "TYPED_PAYLOAD",
-    }
-    .entries(symbol);
-
-    assert_eq!(
-        entries,
-        [
-            AbsentMetadataEntry::new(false, "NAMES", "variant names"),
-            AbsentMetadataEntry::new(true, "OWNER", "variant owner"),
-            AbsentMetadataEntry::new(true, "PAYLOAD_COUNT", "variant payload count"),
-            AbsentMetadataEntry::new(true, "PAYLOAD_TYPE", "variant payload type"),
-            AbsentMetadataEntry::new(true, "TYPED_PAYLOAD", "typed variant payload type"),
-        ]
-    );
-}
-
-#[test]
-fn variant_absence_validation_uses_module_resolver_codes() {
-    let validation = VariantAbsenceValidation::module_resolver_codes();
-
-    assert_eq!(validation.names_code, "E0273");
-    assert_eq!(validation.owner_code, "E0274");
-    assert_eq!(validation.payload_count_code, "E0275");
-    assert_eq!(validation.payload_type_name_code, "E0276");
-    assert_eq!(validation.payload_type_code, "E0375");
-}
-
-#[test]
-fn variant_absence_validation_uses_import_resolver_codes() {
-    let validation = VariantAbsenceValidation::import_resolver_codes();
-
-    assert_eq!(validation.names_code, "E0289");
-    assert_eq!(validation.owner_code, "E0290");
-    assert_eq!(validation.payload_count_code, "E0291");
-    assert_eq!(validation.payload_type_name_code, "E0292");
-    assert_eq!(validation.payload_type_code, "E0366");
-}
-
-#[test]
-fn variant_absence_validation_uses_local_resolver_codes() {
-    let validation = VariantAbsenceValidation::local_resolver_codes();
-
-    assert_eq!(validation.names_code, "E0257");
-    assert_eq!(validation.owner_code, "E0258");
-    assert_eq!(validation.payload_count_code, "E0259");
-    assert_eq!(validation.payload_type_name_code, "E0260");
-    assert_eq!(validation.payload_type_code, "E0384");
-}
-
-#[test]
-fn variant_absence_validation_uses_type_like_resolver_codes() {
-    let validation = VariantAbsenceValidation::type_like_resolver_codes();
-
-    assert_eq!(validation.names_code, "E0315");
-    assert_eq!(validation.owner_code, "E0316");
-    assert_eq!(validation.payload_count_code, "E0317");
-    assert_eq!(validation.payload_type_name_code, "E0318");
-    assert_eq!(validation.payload_type_code, "E0397");
-}
-
-#[test]
-fn variant_absence_validation_uses_behavior_resolver_codes() {
-    let validation = VariantAbsenceValidation::behavior_resolver_codes();
-
-    assert_eq!(validation.names_code, "E0323");
-    assert_eq!(validation.owner_code, "E0324");
-    assert_eq!(validation.payload_count_code, "E0325");
-    assert_eq!(validation.payload_type_name_code, "E0326");
-    assert_eq!(validation.payload_type_code, "E0400");
-}
-
-#[test]
-fn variant_absence_validation_uses_value_resolver_codes() {
-    let validation = VariantAbsenceValidation::value_resolver_codes();
-
-    assert_eq!(validation.names_code, "E0300");
-    assert_eq!(validation.owner_code, "E0301");
-    assert_eq!(validation.payload_count_code, "E0302");
-    assert_eq!(validation.payload_type_name_code, "E0303");
-    assert_eq!(validation.payload_type_code, "E0404");
-}
-
-#[test]
-fn mutability_absence_validation_builds_entries() {
-    let program = parse_program(
-        r#"
-main = (mut input: i32) i32 {
-    value ::= input
-    value
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let symbol = symbols
-        .lookup_scoped(Namespace::Local, "input")
-        .expect("local symbol");
-    let entries = MutabilityAbsenceValidation { code: "MUTABLE" }.entries(symbol);
-
-    assert_eq!(
-        entries,
-        [AbsentMetadataEntry::new(true, "MUTABLE", "mutability")]
-    );
-}
-
-#[test]
-fn mutability_absence_validation_uses_module_resolver_code() {
-    let validation = MutabilityAbsenceValidation::module_resolver_code();
-
-    assert_eq!(validation.code, "E0345");
-}
-
-#[test]
-fn mutability_absence_validation_uses_import_resolver_code() {
-    let validation = MutabilityAbsenceValidation::import_resolver_code();
-
-    assert_eq!(validation.code, "E0344");
-}
-
-#[test]
-fn mutability_absence_validation_uses_type_like_resolver_code() {
-    let validation = MutabilityAbsenceValidation::type_like_resolver_code();
-
-    assert_eq!(validation.code, "E0314");
-}
-
-#[test]
-fn mutability_absence_validation_uses_variant_resolver_code() {
-    let validation = MutabilityAbsenceValidation::variant_resolver_code();
-
-    assert_eq!(validation.code, "E0343");
-}
-
-#[test]
-fn mutability_absence_validation_uses_value_resolver_code() {
-    let validation = MutabilityAbsenceValidation::value_resolver_code();
-
-    assert_eq!(validation.code, "E0308");
-}
-
-#[test]
-fn mutability_validation_formats_actual_and_expected() {
-    let validation = MutabilityValidation { code: "MUTABLE" };
-
-    assert_eq!(validation.code, "MUTABLE");
-    assert_eq!(
-        validation.display(Some(false), true),
-        ("immutable", "mutable")
-    );
-    assert_eq!(validation.display(None, false), ("unknown", "immutable"));
-    assert_eq!(
-        validation.message("local", "value", Some(false), true),
-        "resolver local symbol 'value' has mutability immutable, expected mutable"
-    );
-}
-
-#[test]
-fn mutability_validation_uses_resolver_code() {
-    let validation = MutabilityValidation::resolver_code();
-
-    assert_eq!(validation.code, "E0231");
-}
-
-#[test]
-fn visibility_validation_formats_actual_and_expected() {
-    let validation = VisibilityValidation { code: "VISIBLE" };
-
-    assert_eq!(validation.code, "VISIBLE");
-    assert_eq!(validation.display(true, false), ("public", "private"));
-    assert_eq!(validation.display(false, true), ("private", "public"));
-    assert_eq!(
-        validation.message("import", "io", true, false),
-        "resolver import symbol 'io' has visibility public, expected private"
-    );
-}
-
-#[test]
-fn visibility_validation_uses_local_resolver_code() {
-    let validation = VisibilityValidation::local_resolver_code();
-
-    assert_eq!(validation.code, "E0247");
-}
-
-#[test]
-fn visibility_validation_uses_module_resolver_code() {
-    let validation = VisibilityValidation::module_resolver_code();
-
-    assert_eq!(validation.code, "E0229");
-}
-
-#[test]
-fn visibility_validation_uses_import_resolver_code() {
-    let validation = VisibilityValidation::import_resolver_code();
-
-    assert_eq!(validation.code, "E0245");
-}
-
-#[test]
-fn visibility_validation_uses_type_like_resolver_code() {
-    let validation = VisibilityValidation::type_like_resolver_code();
-
-    assert_eq!(validation.code, "E0225");
-}
-
-#[test]
-fn visibility_validation_uses_variant_resolver_code() {
-    let validation = VisibilityValidation::variant_resolver_code();
-
-    assert_eq!(validation.code, "E0226");
-}
-
-#[test]
-fn visibility_validation_uses_value_resolver_code() {
-    let validation = VisibilityValidation::value_resolver_code();
-
-    assert_eq!(validation.code, "E0224");
-}
-
-#[test]
-fn resolver_symbol_presence_validation_formats_messages() {
-    let extra = ResolverSymbolPresenceValidation {
-        code: "EXTRA",
-        presence: ResolverSymbolPresence::Extra,
-    };
-    let missing = ResolverSymbolPresenceValidation {
-        code: "MISSING",
-        presence: ResolverSymbolPresence::Missing,
-    };
-
-    assert_eq!(extra.code, "EXTRA");
-    assert_eq!(
-        extra.message("value", "main"),
-        "resolver symbol table has extra value symbol 'main'"
-    );
-    assert_eq!(missing.code, "MISSING");
-    assert_eq!(
-        missing.message("local", "value"),
-        "resolver symbol table missing local symbol 'value'"
-    );
-}
-
-#[test]
-fn resolver_symbol_presence_validation_uses_resolver_codes() {
-    let missing = ResolverSymbolPresenceValidation::missing_resolver_code();
-    let missing_local = ResolverSymbolPresenceValidation::missing_local_resolver_code();
-    let extra_declaration = ResolverSymbolPresenceValidation::extra_declaration_resolver_code();
-    let extra_local = ResolverSymbolPresenceValidation::extra_local_resolver_code();
-
-    assert_eq!(missing.code, "E0210");
-    assert!(matches!(missing.presence, ResolverSymbolPresence::Missing));
-    assert_eq!(missing_local.code, "E0228");
-    assert!(matches!(
-        missing_local.presence,
-        ResolverSymbolPresence::Missing
-    ));
-    assert_eq!(extra_declaration.code, "E0243");
-    assert!(matches!(
-        extra_declaration.presence,
-        ResolverSymbolPresence::Extra
-    ));
-    assert_eq!(extra_local.code, "E0244");
-    assert!(matches!(
-        extra_local.presence,
-        ResolverSymbolPresence::Extra
-    ));
-}
-
-#[test]
-fn resolver_symbol_presence_validation_pushes_diagnostic() {
-    let mut tc = TypeChecker::new();
-
-    tc.validate_resolver_symbol_presence(
-        "value",
-        "main",
-        ResolverSymbolPresenceValidation {
-            code: "EXTRA",
-            presence: ResolverSymbolPresence::Extra,
-        },
-        Span::dummy(),
-    );
-
-    assert_eq!(tc.diagnostics.len(), 1);
-    assert_eq!(tc.diagnostics[0].code, "EXTRA");
-    assert_eq!(
-        tc.diagnostics[0].message,
-        "resolver symbol table has extra value symbol 'main'"
-    );
-}
-
-#[test]
-fn source_absence_validation_builds_source_validation() {
-    let validation = SourceAbsenceValidation { code: "SOURCE" }.source_validation();
-
-    assert_eq!(validation.code, "SOURCE");
-    assert_eq!(validation.actual_missing, "none");
-    assert_eq!(validation.expected_missing, "none");
-    assert!(!validation.quote_expected);
-}
-
-#[test]
-fn source_absence_validation_uses_type_like_resolver_code() {
-    let validation = SourceAbsenceValidation::type_like_resolver_code();
-
-    assert_eq!(validation.code, "E0309");
-}
-
-#[test]
-fn source_absence_validation_uses_variant_resolver_code() {
-    let validation = SourceAbsenceValidation::variant_resolver_code();
-
-    assert_eq!(validation.code, "E0329");
-}
-
-#[test]
-fn source_absence_validation_uses_value_resolver_code() {
-    let validation = SourceAbsenceValidation::value_resolver_code();
-
-    assert_eq!(validation.code, "E0297");
-}
-
-#[test]
-fn source_validation_formats_message() {
-    let quoted = SourceValidation {
-        code: "SOURCE",
-        actual_missing: "unknown",
-        expected_missing: "none",
-        quote_expected: true,
-    };
-    let unquoted = SourceValidation {
-        code: "SOURCE",
-        actual_missing: "none",
-        expected_missing: "none",
-        quote_expected: false,
-    };
-
-    assert_eq!(
-        quoted.message("import", "io", Some("other"), Some("std")),
-        "resolver import symbol 'io' has source 'other', expected 'std'"
-    );
-    assert_eq!(
-        unquoted.message("value", "main", Some("std"), None),
-        "resolver value symbol 'main' has source 'std', expected none"
-    );
-}
-
-#[test]
-fn source_validation_uses_resolver_codes() {
-    let module = SourceValidation::module_resolver_code();
-    let stripped_import = SourceValidation::stripped_import_resolver_code();
-    let import = SourceValidation::import_resolver_code();
-    let local = SourceValidation::local_resolver_code();
-
-    assert_eq!(module.code, "E0230");
-    assert_eq!(module.actual_missing, "none");
-    assert_eq!(module.expected_missing, "none");
-    assert!(!module.quote_expected);
-    assert_eq!(stripped_import.code, "E0246");
-    assert_eq!(stripped_import.actual_missing, "unknown");
-    assert_eq!(stripped_import.expected_missing, "a module source");
-    assert!(!stripped_import.quote_expected);
-    assert_eq!(import.code, "E0227");
-    assert_eq!(import.actual_missing, "unknown");
-    assert_eq!(import.expected_missing, "none");
-    assert!(import.quote_expected);
-    assert_eq!(local.code, "E0248");
-    assert_eq!(local.actual_missing, "none");
-    assert_eq!(local.expected_missing, "none");
-    assert!(!local.quote_expected);
-}
-
-#[test]
-fn absent_metadata_entry_formats_message() {
-    let entry = AbsentMetadataEntry {
-        present: true,
-        code: "ABSENT",
-        label: "parameter count",
-    };
-
-    assert_eq!(entry.code, "ABSENT");
-    assert_eq!(
-        entry.message("value", "main"),
-        "resolver value symbol 'main' has parameter count metadata, expected none"
-    );
-}
-
-#[test]
-fn resolver_named_list_display_formats_known_and_missing_items() {
-    let fields = vec![("value".to_string(), "i32".to_string())];
-    assert_eq!(
-        format_resolver_named_list(Some(&fields), |ty: &String| ty.clone()),
-        "(value: i32)"
-    );
-    assert_eq!(
-        format_resolver_named_list::<String>(None, |ty: &String| ty.clone()),
-        "unknown"
-    );
-}
-
-#[test]
-fn check_program_rejects_self_type_outside_method_or_behavior() {
-    let program = parse_program(
-        r#"
-main = (value: Self) i32 { 0 }
-"#,
-    );
-    let mut tc = TypeChecker::new();
-
-    let err = tc
-        .check_program(&program)
-        .expect_err("Self should require a method or behavior context");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("Self type is only valid")),
-        "expected invalid Self type diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn self_type_context_validation_tasks_collect_declarations() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-Option<T>: Some(T), None
-
-Json: behavior {
-    encode: (Self) str
-}
-
-Pretty.extends(Json)
-
-make = () Point { Point { x: 1 } }
-
-Point.get = (self: Point) i32 { self.x }
-
-Point.impl = {
-    x_value = (self: Point) i32 { self.x }
-}
-
-Point.requires(Json)
-
-result := 1
-"#,
-    );
-
-    let tasks = TypeChecker::collect_self_type_context_validation_tasks(&program.declarations);
-
-    assert_eq!(tasks.len(), 9);
-    assert!(matches!(
-        tasks[0],
-        SelfTypeContextValidationTask::Struct { .. }
-    ));
-    assert!(matches!(
-        tasks[1],
-        SelfTypeContextValidationTask::Enum { .. }
-    ));
-    assert!(matches!(
-        tasks[2],
-        SelfTypeContextValidationTask::Behavior { .. }
-    ));
-    assert!(matches!(
-        tasks[3],
-        SelfTypeContextValidationTask::BehaviorExtends { .. }
-    ));
-    assert!(matches!(
-        tasks[4],
-        SelfTypeContextValidationTask::Function { .. }
-    ));
-    assert!(matches!(
-        tasks[5],
-        SelfTypeContextValidationTask::Method { .. }
-    ));
-    assert!(matches!(
-        tasks[6],
-        SelfTypeContextValidationTask::ImplBlock { .. }
-    ));
-    assert!(matches!(
-        tasks[7],
-        SelfTypeContextValidationTask::Requires { .. }
-    ));
-    assert!(matches!(
-        tasks[8],
-        SelfTypeContextValidationTask::TopLevelExpr { .. }
-    ));
-}
-
-#[test]
-fn ast_precollection_validation_tasks_collect_self_and_extends_work() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    encode: (Self) str
-}
-
-Pretty.extends(Json)
-
-Point.impl = {
-    x_value = (self: Point) i32 { self.x }
-}
-
-result := 1
-"#,
-    );
-
-    let tasks = TypeChecker::collect_ast_precollection_validation_tasks(&program.declarations);
-
-    assert_eq!(tasks.self_type_contexts.len(), 5);
-    assert_eq!(tasks.behavior_associations.extends.len(), 1);
-    assert_eq!(tasks.behavior_associations.extends[0].behavior, "Pretty");
-    assert_eq!(tasks.behavior_associations.extends[0].parent, "Json");
-    assert!(tasks.behavior_associations.impls.is_empty());
-    assert!(tasks.behavior_associations.requires.is_empty());
-}
-
-#[test]
-fn ast_declaration_collection_tasks_include_precollection_validation_work() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    encode: (Self) str
-}
-
-Pretty.extends(Json)
-
-make = () Point { Point { x: 1 } }
-"#,
-    );
-
-    let tasks = TypeChecker::collect_ast_declaration_collection_tasks(&program.declarations);
-
-    assert_eq!(tasks.types.len(), 1);
-    assert_eq!(tasks.behaviors.len(), 1);
-    assert_eq!(tasks.callable.len(), 1);
-    assert_eq!(tasks.precollection_validations.self_type_contexts.len(), 4);
-    assert_eq!(
-        tasks
-            .precollection_validations
-            .behavior_associations
-            .extends
-            .len(),
-        1
-    );
-}
-
-#[test]
-fn check_program_rejects_unknown_type_references() {
-    let program = parse_program(
-        r#"
-main = (value: Missing, items: Bag<i32>) i32 { 0 }
-"#,
-    );
-    let mut tc = TypeChecker::new();
-
-    let err = tc
-        .check_program(&program)
-        .expect_err("unknown type reference should fail");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("unknown type symbol 'Missing'")),
-        "expected unknown type diagnostic, got {err:?}"
-    );
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("unknown type symbol 'Bag'")),
-        "expected unknown generic type diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn check_program_rejects_unknown_type_references_in_struct_field_defaults() {
-    let program = parse_program(
-        r#"
-Box<T>: {
-    value: T = {
-        same: Missing = 1
-        same
-    }
-}
-"#,
-    );
-    let mut tc = TypeChecker::new();
-
-    let err = tc
-        .check_program(&program)
-        .expect_err("unknown struct field default type reference should fail");
-
-    assert!(
-        err.iter()
-            .any(|d| d.message.contains("unknown type symbol 'Missing'")),
-        "expected unknown field default type diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn check_program_rejects_struct_field_default_type_mismatch() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 = "bad" }
-"#,
-    );
-    let mut tc = TypeChecker::new();
-
-    let err = tc
-        .check_program(&program)
-        .expect_err("struct field default type mismatch should fail");
-
-    assert!(
-        err.iter().any(|d| d
-            .message
-            .contains("field `x` default expects `i32`, found `str`")),
-        "expected field default type mismatch diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn ast_struct_field_default_validation_tasks_collect_structs() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 = 1 }
-Box<T>: { value: T }
-"#,
-    );
-
-    let tasks =
-        TypeChecker::collect_ast_struct_field_default_validation_tasks(&program.declarations);
-
-    assert_eq!(tasks.len(), 2);
-    assert_eq!(tasks[0].type_params.len(), 0);
-    assert_eq!(tasks[0].fields.len(), 1);
-    assert_eq!(tasks[1].type_params.len(), 1);
-    assert_eq!(tasks[1].fields.len(), 1);
-}
-
-#[test]
-fn resolver_type_declaration_metadata_tasks_collect_only_type_work() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 = 1 }
-Option<T>: Some(T), None
-
-main = () i32 { 1 }
-"#,
-    );
-
-    let tasks =
-        TypeChecker::collect_resolver_type_declaration_metadata_tasks(&program.declarations);
-
-    assert_eq!(tasks.len(), 2);
-    assert!(matches!(
-        tasks[0],
-        ResolverTypeDeclarationMetadataTask::Struct { name: "Point", .. }
-    ));
-    assert!(matches!(
-        tasks[1],
-        ResolverTypeDeclarationMetadataTask::Enum { name: "Option", .. }
-    ));
-}
-
-#[test]
-fn resolver_type_replay_task_helper_pushes_metadata_and_type_refs_together() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 = 1 }
-"#,
-    );
-    let mut type_tasks = Vec::new();
-    let mut type_reference_tasks = Vec::new();
-
-    let handled = TypeChecker::push_resolver_type_replay_tasks(
-        &program.declarations[0],
-        &mut type_tasks,
-        &mut type_reference_tasks,
-    );
-
-    assert!(handled);
-    assert!(matches!(
-        type_tasks.as_slice(),
-        [ResolverTypeDeclarationMetadataTask::Struct { name: "Point", .. }]
-    ));
-    assert!(matches!(
-        type_reference_tasks.as_slice(),
-        [ResolverTypeReferenceValidationTask::Struct { name: "Point", .. }]
-    ));
-}
-
-#[test]
-fn resolver_behavior_declaration_metadata_tasks_collect_only_behavior_work() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-main = () i32 { 1 }
-"#,
-    );
-
-    let tasks =
-        TypeChecker::collect_resolver_behavior_declaration_metadata_tasks(&program.declarations);
-
-    assert_eq!(tasks.len(), 1);
-    assert_eq!(tasks[0].name, "Json");
-}
-
-#[test]
-fn resolver_behavior_replay_task_helper_pushes_metadata_and_type_refs_together() {
-    let program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-"#,
-    );
-    let mut behavior_tasks = Vec::new();
-    let mut type_reference_tasks = Vec::new();
-
-    let handled = TypeChecker::push_resolver_behavior_replay_tasks(
-        &program.declarations[0],
-        &mut behavior_tasks,
-        &mut type_reference_tasks,
-    );
-
-    assert!(handled);
-    assert_eq!(behavior_tasks.len(), 1);
-    assert_eq!(behavior_tasks[0].name, "Json");
-    assert!(matches!(
-        type_reference_tasks.as_slice(),
-        [ResolverTypeReferenceValidationTask::Behavior { name: "Json", .. }]
-    ));
-}
-
-#[test]
-fn resolver_behavior_impl_replay_task_helper_pushes_metadata_and_type_refs_together() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    encode: (Self) str
-}
-
-Point.implements(Json) {
-    encode = (self: Point) str { "point" }
-}
-"#,
-    );
-    let mut behavior_impl_tasks = Vec::new();
-    let mut type_reference_tasks = Vec::new();
-
-    let handled = TypeChecker::push_resolver_behavior_impl_replay_tasks(
-        &program.declarations[2],
-        &mut behavior_impl_tasks,
-        &mut type_reference_tasks,
-    );
-
-    assert!(handled);
-    assert_eq!(behavior_impl_tasks.len(), 1);
-    assert_eq!(behavior_impl_tasks[0].ast_type_name, "Point");
-    assert_eq!(behavior_impl_tasks[0].behavior, "Json");
-    assert!(matches!(
-        type_reference_tasks.as_slice(),
-        [ResolverTypeReferenceValidationTask::ImplBlock {
-            type_name: "Point",
-            ..
-        }]
-    ));
-}
-
-#[test]
-fn resolver_callable_declaration_metadata_tasks_collect_callable_work() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-make = () Point { Point { x: 1 } }
-
-Point.get = (self: Point) i32 { self.x }
-
-Point.impl = {
-    plus = (self: Point, other: Point) i32 { self.x + other.x }
-}
-"#,
-    );
-
-    let tasks =
-        TypeChecker::collect_resolver_callable_declaration_metadata_tasks(&program.declarations);
-
-    assert_eq!(tasks.len(), 3);
-    assert!(matches!(
-        tasks[0],
-        ResolverCallableDeclarationMetadataTask::Function { name: "make", .. }
-    ));
-    assert!(matches!(
-        tasks[1],
-        ResolverCallableDeclarationMetadataTask::Method {
-            type_name: "Point",
-            method_name: "get",
-            ..
-        }
-    ));
-    assert!(matches!(
-        tasks[2],
-        ResolverCallableDeclarationMetadataTask::TypeImpl {
-            type_name: "Point",
-            ..
-        }
-    ));
-}
-
-#[test]
-fn resolver_callable_replay_task_helper_pushes_metadata_and_type_refs_together() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-make = () Point { Point { x: 1 } }
-"#,
-    );
-    let mut callable_tasks = Vec::new();
-    let mut type_reference_tasks = Vec::new();
-
-    let handled = TypeChecker::push_resolver_callable_replay_tasks(
-        &program.declarations[1],
-        &mut callable_tasks,
-        &mut type_reference_tasks,
-    );
-
-    assert!(handled);
-    assert!(matches!(
-        callable_tasks.as_slice(),
-        [ResolverCallableDeclarationMetadataTask::Function { name: "make", .. }]
-    ));
-    assert!(matches!(
-        type_reference_tasks.as_slice(),
-        [ResolverTypeReferenceValidationTask::Function { name: "make", .. }]
-    ));
-}
-
-#[test]
-fn resolver_behavior_impl_block_declaration_tasks_collect_only_behavior_impls() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    encode: (Self) str
-}
-
-Point.impl = {
-    x_value = (value: Point) i32 { value.x }
-}
-
-Point.implements(Json) {
-    encode = (value: Point) str { "point" }
-}
-"#,
-    );
-
-    let tasks =
-        TypeChecker::collect_resolver_behavior_impl_block_declaration_tasks(&program.declarations);
-
-    assert_eq!(tasks.len(), 1);
-    assert_eq!(tasks[0].ast_type_name, "Point");
-    assert_eq!(tasks[0].behavior, "Json");
-    assert_eq!(tasks[0].methods.len(), 1);
-}
-
-#[test]
-fn ast_type_reference_validation_tasks_collect_declarations() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-Option<T>: Some(T), None
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-make = () Point { Point { x: 1 } }
-
-Point.get = (self: Point) i32 { self.x }
-
-Point.impl = {
-    x_value = (self: Point) i32 { self.x }
-}
-
-result := make()
-"#,
-    );
-
-    let tasks = TypeChecker::collect_ast_type_reference_validation_tasks(&program.declarations);
-
-    assert_eq!(tasks.len(), 7);
-    assert!(matches!(
-        tasks[0],
-        AstTypeReferenceValidationTask::Struct { .. }
-    ));
-    assert!(matches!(
-        tasks[1],
-        AstTypeReferenceValidationTask::Enum { .. }
-    ));
-    assert!(matches!(
-        tasks[2],
-        AstTypeReferenceValidationTask::Behavior { .. }
-    ));
-    assert!(matches!(
-        tasks[3],
-        AstTypeReferenceValidationTask::Function { .. }
-    ));
-    assert!(matches!(
-        tasks[4],
-        AstTypeReferenceValidationTask::Method { .. }
-    ));
-    assert!(matches!(
-        tasks[5],
-        AstTypeReferenceValidationTask::ImplBlock { .. }
-    ));
-    assert!(matches!(
-        tasks[6],
-        AstTypeReferenceValidationTask::TopLevelExpr { .. }
-    ));
-}
-
-#[test]
-fn ast_declaration_validation_tasks_collect_semantic_validation_work() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 = 1 }
-Option<T>: Some(T), None
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Point.implements(Json<str>) {
-    encode = (self: Point) str { "point" }
-}
-
-Point.requires(Json<str>)
-JsonString.extends(Json<str>)
-
-main = () i32 { 1 }
-"#,
-    );
-
-    let tasks = TypeChecker::collect_ast_declaration_validation_tasks(&program.declarations);
-
-    assert_eq!(tasks.behavior_associations.extends.len(), 1);
-    assert_eq!(tasks.behavior_associations.impls.len(), 1);
-    assert_eq!(tasks.behavior_associations.requires.len(), 1);
-    assert_eq!(tasks.type_references.len(), 5);
-    assert_eq!(tasks.struct_field_defaults.len(), 1);
-}
+mod resolver_absence;
+
+mod core_semantics;
+mod declaration_validation;
+mod generic_behaviors;
 
 #[test]
 fn scope_variable_lookup() {
@@ -10012,1278 +8772,3 @@ Option: Some(i32), None
         "expected resolver enum variant owner diagnostic, got {err:?}"
     );
 }
-
-#[test]
-fn binary_op_types() {
-    let tc = TypeChecker::new();
-    assert_eq!(
-        tc.check_binary_op(BinaryOp::Add, &Type::I32, &Type::I32, &Span::dummy())
-            .unwrap(),
-        Type::I32
-    );
-    assert_eq!(
-        tc.check_binary_op(BinaryOp::Eq, &Type::I32, &Type::I32, &Span::dummy())
-            .unwrap(),
-        Type::Bool
-    );
-    assert_eq!(
-        tc.check_binary_op(BinaryOp::And, &Type::Bool, &Type::Bool, &Span::dummy())
-            .unwrap(),
-        Type::Bool
-    );
-}
-
-#[test]
-fn binary_op_type_mismatch() {
-    let tc = TypeChecker::new();
-    // Arithmetic on non-numeric type
-    assert!(tc
-        .check_binary_op(BinaryOp::Add, &Type::I32, &Type::Str, &Span::dummy())
-        .is_err());
-    assert!(tc
-        .check_binary_op(BinaryOp::Add, &Type::Bool, &Type::I32, &Span::dummy())
-        .is_err());
-    // Logical op on non-bool
-    assert!(tc
-        .check_binary_op(BinaryOp::And, &Type::I32, &Type::Bool, &Span::dummy())
-        .is_err());
-    // Unknown is permissive (error recovery)
-    assert!(tc
-        .check_binary_op(BinaryOp::Add, &Type::Unknown, &Type::Str, &Span::dummy())
-        .is_ok());
-}
-
-#[test]
-fn binary_op_mixed_numeric_width_requires_cast() {
-    let tc = TypeChecker::new();
-    let err = tc
-        .check_binary_op(BinaryOp::Add, &Type::I32, &Type::I64, &Span::dummy())
-        .expect_err("mixed integer arithmetic should fail");
-    assert!(
-        err.message
-            .contains("arithmetic operands must have the same type"),
-        "expected mixed numeric diagnostic, got {err:?}"
-    );
-
-    let err = tc
-        .check_binary_op(BinaryOp::Mul, &Type::F32, &Type::F64, &Span::dummy())
-        .expect_err("mixed float arithmetic should fail");
-    assert!(
-        err.message
-            .contains("arithmetic operands must have the same type"),
-        "expected mixed numeric diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn unknown_function_error() {
-    use crate::ast::{Expression, Program};
-    let mut tc = TypeChecker::new();
-    let program = Program {
-        declarations: vec![Declaration::Function {
-            name: "main".into(),
-            type_params: Vec::new(),
-            params: Vec::new(),
-            return_type: Some(AstType::Void),
-            body: Expression::Block {
-                statements: vec![ast::Statement::Expression {
-                    expr: Expression::FunctionCall {
-                        name: "nonexistent".into(),
-                        module: None,
-                        type_args: Vec::new(),
-                        args: Vec::new(),
-                        span: Span::dummy(),
-                    },
-                    span: Span::dummy(),
-                }],
-                expr: None,
-                span: Span::dummy(),
-            },
-            public: false,
-            span: Span::dummy(),
-        }],
-        file_id: 0,
-    };
-    let result = tc.check_program(&program);
-    assert!(result.is_err());
-    let errors = result.unwrap_err();
-    assert!(errors
-        .iter()
-        .any(|d| d.message.contains("undefined function")));
-}
-
-#[test]
-fn return_type_mismatch_error() {
-    use crate::ast::{Expression, Program};
-    let mut tc = TypeChecker::new();
-    let program = Program {
-        declarations: vec![Declaration::Function {
-            name: "foo".into(),
-            type_params: Vec::new(),
-            params: Vec::new(),
-            return_type: Some(AstType::I32),
-            body: Expression::Block {
-                statements: Vec::new(),
-                expr: Some(Box::new(Expression::Return {
-                    value: Some(Box::new(Expression::StringLiteral {
-                        value: "hello".into(),
-                        span: Span::dummy(),
-                    })),
-                    span: Span::dummy(),
-                })),
-                span: Span::dummy(),
-            },
-            public: false,
-            span: Span::dummy(),
-        }],
-        file_id: 0,
-    };
-    let result = tc.check_program(&program);
-    assert!(result.is_err());
-    let errors = result.unwrap_err();
-    assert!(errors
-        .iter()
-        .any(|d| d.message.contains("return type mismatch")));
-}
-
-#[test]
-fn function_call_wrong_arity_is_error() {
-    use crate::ast::{Expression, Program};
-    let mut tc = TypeChecker::new();
-    let program = Program {
-        declarations: vec![
-            Declaration::Function {
-                name: "add".into(),
-                type_params: Vec::new(),
-                params: vec![
-                    ast::Param {
-                        name: "a".into(),
-                        ty: AstType::I32,
-                        mutable: false,
-                        span: Span::dummy(),
-                    },
-                    ast::Param {
-                        name: "b".into(),
-                        ty: AstType::I32,
-                        mutable: false,
-                        span: Span::dummy(),
-                    },
-                ],
-                return_type: Some(AstType::I32),
-                body: Expression::Block {
-                    statements: Vec::new(),
-                    expr: Some(Box::new(Expression::Return {
-                        value: Some(Box::new(Expression::Identifier {
-                            name: "a".into(),
-                            span: Span::dummy(),
-                        })),
-                        span: Span::dummy(),
-                    })),
-                    span: Span::dummy(),
-                },
-                public: false,
-                span: Span::dummy(),
-            },
-            Declaration::Function {
-                name: "main".into(),
-                type_params: Vec::new(),
-                params: Vec::new(),
-                return_type: Some(AstType::Void),
-                body: Expression::Block {
-                    statements: vec![ast::Statement::Expression {
-                        expr: Expression::FunctionCall {
-                            name: "add".into(),
-                            module: None,
-                            type_args: Vec::new(),
-                            args: vec![Expression::IntLiteral {
-                                value: 1,
-                                span: Span::dummy(),
-                            }],
-                            span: Span::dummy(),
-                        },
-                        span: Span::dummy(),
-                    }],
-                    expr: None,
-                    span: Span::dummy(),
-                },
-                public: false,
-                span: Span::dummy(),
-            },
-        ],
-        file_id: 0,
-    };
-
-    let errors = tc
-        .check_program(&program)
-        .expect_err("wrong arity should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("function `add` expects 2 arguments, found 1")),
-        "expected arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn function_call_argument_type_mismatch_is_error() {
-    use crate::ast::{Expression, Program};
-    let mut tc = TypeChecker::new();
-    let program = Program {
-        declarations: vec![
-            Declaration::Function {
-                name: "takes_i32".into(),
-                type_params: Vec::new(),
-                params: vec![ast::Param {
-                    name: "value".into(),
-                    ty: AstType::I32,
-                    mutable: false,
-                    span: Span::dummy(),
-                }],
-                return_type: Some(AstType::Void),
-                body: Expression::Block {
-                    statements: Vec::new(),
-                    expr: None,
-                    span: Span::dummy(),
-                },
-                public: false,
-                span: Span::dummy(),
-            },
-            Declaration::Function {
-                name: "main".into(),
-                type_params: Vec::new(),
-                params: Vec::new(),
-                return_type: Some(AstType::Void),
-                body: Expression::Block {
-                    statements: vec![ast::Statement::Expression {
-                        expr: Expression::FunctionCall {
-                            name: "takes_i32".into(),
-                            module: None,
-                            type_args: Vec::new(),
-                            args: vec![Expression::StringLiteral {
-                                value: "bad".into(),
-                                span: Span::dummy(),
-                            }],
-                            span: Span::dummy(),
-                        },
-                        span: Span::dummy(),
-                    }],
-                    expr: None,
-                    span: Span::dummy(),
-                },
-                public: false,
-                span: Span::dummy(),
-            },
-        ],
-        file_id: 0,
-    };
-
-    let errors = tc
-        .check_program(&program)
-        .expect_err("argument type mismatch should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("argument 1 for `takes_i32` expects `i32`, found `str`")),
-        "expected argument type diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn struct_literal_missing_field_is_error() {
-    use crate::ast::declarations::StructField;
-    use crate::ast::{Expression, Program, Statement};
-    let mut tc = TypeChecker::new();
-    let program = Program {
-        declarations: vec![
-            Declaration::Struct {
-                name: "Point".into(),
-                type_params: Vec::new(),
-                fields: vec![
-                    StructField {
-                        name: "x".into(),
-                        ty: AstType::I32,
-                        default: None,
-                        mutable: false,
-                        span: Span::dummy(),
-                    },
-                    StructField {
-                        name: "y".into(),
-                        ty: AstType::I32,
-                        default: None,
-                        mutable: false,
-                        span: Span::dummy(),
-                    },
-                ],
-                public: false,
-                span: Span::dummy(),
-            },
-            Declaration::Function {
-                name: "main".into(),
-                type_params: Vec::new(),
-                params: Vec::new(),
-                return_type: Some(AstType::Void),
-                body: Expression::Block {
-                    statements: vec![Statement::VarDecl {
-                        name: "p".into(),
-                        ty: None,
-                        value: Expression::StructLiteral {
-                            name: "Point".into(),
-                            type_args: Vec::new(),
-                            fields: vec![(
-                                "x".into(),
-                                Expression::IntLiteral {
-                                    value: 1,
-                                    span: Span::dummy(),
-                                },
-                            )],
-                            span: Span::dummy(),
-                        },
-                        mutable: false,
-                        constant: false,
-                        span: Span::dummy(),
-                    }],
-                    expr: None,
-                    span: Span::dummy(),
-                },
-                public: false,
-                span: Span::dummy(),
-            },
-        ],
-        file_id: 0,
-    };
-
-    let errors = tc
-        .check_program(&program)
-        .expect_err("missing struct field should fail");
-    assert!(
-        errors
-            .iter()
-            .any(|d| d.message.contains("missing field `y` for struct `Point`")),
-        "expected missing field diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn struct_literal_uses_default_for_omitted_field() {
-    use crate::ast::declarations::StructField;
-    use crate::ast::typed::{TypedExprKind, TypedStatementKind};
-    use crate::ast::{Expression, Program, Statement};
-    let mut tc = TypeChecker::new();
-    let program = Program {
-        declarations: vec![
-            Declaration::Struct {
-                name: "Point".into(),
-                type_params: Vec::new(),
-                fields: vec![
-                    StructField {
-                        name: "x".into(),
-                        ty: AstType::I32,
-                        default: None,
-                        mutable: false,
-                        span: Span::dummy(),
-                    },
-                    StructField {
-                        name: "y".into(),
-                        ty: AstType::I32,
-                        default: Some(Expression::IntLiteral {
-                            value: 2,
-                            span: Span::dummy(),
-                        }),
-                        mutable: false,
-                        span: Span::dummy(),
-                    },
-                ],
-                public: false,
-                span: Span::dummy(),
-            },
-            Declaration::Function {
-                name: "main".into(),
-                type_params: Vec::new(),
-                params: Vec::new(),
-                return_type: Some(AstType::Void),
-                body: Expression::Block {
-                    statements: vec![Statement::VarDecl {
-                        name: "p".into(),
-                        ty: None,
-                        value: Expression::StructLiteral {
-                            name: "Point".into(),
-                            type_args: Vec::new(),
-                            fields: vec![(
-                                "x".into(),
-                                Expression::IntLiteral {
-                                    value: 1,
-                                    span: Span::dummy(),
-                                },
-                            )],
-                            span: Span::dummy(),
-                        },
-                        mutable: false,
-                        constant: false,
-                        span: Span::dummy(),
-                    }],
-                    expr: None,
-                    span: Span::dummy(),
-                },
-                public: false,
-                span: Span::dummy(),
-            },
-        ],
-        file_id: 0,
-    };
-
-    let typed = tc
-        .check_program(&program)
-        .expect("defaulted struct field may be omitted");
-    let TypedStatementKind::VarDecl { value, .. } = &typed.functions[0].body.statements[0].kind
-    else {
-        panic!("expected var decl");
-    };
-    let TypedExprKind::StructLiteral { fields, .. } = &value.kind else {
-        panic!("expected struct literal");
-    };
-    assert_eq!(fields.len(), 2);
-    assert_eq!(fields[1].0, "y");
-    assert!(matches!(fields[1].1.kind, TypedExprKind::IntLiteral(2)));
-}
-
-#[test]
-fn generic_struct_literal_uses_substituted_default_for_omitted_field() {
-    use crate::ast::declarations::{StructField, TypeParam};
-    use crate::ast::typed::{TypedExprKind, TypedStatementKind};
-    use crate::ast::{Expression, Program, Statement};
-    let mut tc = TypeChecker::new();
-    let program = Program {
-        declarations: vec![
-            Declaration::Struct {
-                name: "Box".into(),
-                type_params: vec![TypeParam {
-                    name: "T".into(),
-                    constraint: None,
-                    constraint_type_args: Vec::new(),
-                    span: Span::dummy(),
-                }],
-                fields: vec![StructField {
-                    name: "value".into(),
-                    ty: AstType::Named("T".into()),
-                    default: Some(Expression::Block {
-                        statements: vec![Statement::VarDecl {
-                            name: "same".into(),
-                            ty: Some(AstType::Named("T".into())),
-                            value: Expression::StringLiteral {
-                                value: "fallback".into(),
-                                span: Span::dummy(),
-                            },
-                            mutable: false,
-                            constant: false,
-                            span: Span::dummy(),
-                        }],
-                        expr: Some(Box::new(Expression::Identifier {
-                            name: "same".into(),
-                            span: Span::dummy(),
-                        })),
-                        span: Span::dummy(),
-                    }),
-                    mutable: false,
-                    span: Span::dummy(),
-                }],
-                public: false,
-                span: Span::dummy(),
-            },
-            Declaration::Function {
-                name: "main".into(),
-                type_params: Vec::new(),
-                params: Vec::new(),
-                return_type: Some(AstType::Void),
-                body: Expression::Block {
-                    statements: vec![Statement::VarDecl {
-                        name: "box".into(),
-                        ty: None,
-                        value: Expression::StructLiteral {
-                            name: "Box".into(),
-                            type_args: vec![AstType::Str],
-                            fields: Vec::new(),
-                            span: Span::dummy(),
-                        },
-                        mutable: false,
-                        constant: false,
-                        span: Span::dummy(),
-                    }],
-                    expr: None,
-                    span: Span::dummy(),
-                },
-                public: false,
-                span: Span::dummy(),
-            },
-        ],
-        file_id: 0,
-    };
-
-    let typed = tc
-        .check_program(&program)
-        .expect("generic defaulted struct field may be omitted");
-    let TypedStatementKind::VarDecl { value, .. } = &typed.functions[0].body.statements[0].kind
-    else {
-        panic!("expected var decl");
-    };
-    let TypedExprKind::StructLiteral { fields, .. } = &value.kind else {
-        panic!("expected struct literal");
-    };
-    assert_eq!(fields.len(), 1);
-    assert_eq!(fields[0].0, "value");
-    assert_eq!(fields[0].1.ty, Type::Str);
-}
-
-#[test]
-fn struct_literal_field_type_mismatch_is_error() {
-    use crate::ast::declarations::StructField;
-    use crate::ast::{Expression, Program, Statement};
-    let mut tc = TypeChecker::new();
-    let program = Program {
-        declarations: vec![
-            Declaration::Struct {
-                name: "Point".into(),
-                type_params: Vec::new(),
-                fields: vec![StructField {
-                    name: "x".into(),
-                    ty: AstType::I32,
-                    default: None,
-                    mutable: false,
-                    span: Span::dummy(),
-                }],
-                public: false,
-                span: Span::dummy(),
-            },
-            Declaration::Function {
-                name: "main".into(),
-                type_params: Vec::new(),
-                params: Vec::new(),
-                return_type: Some(AstType::Void),
-                body: Expression::Block {
-                    statements: vec![Statement::VarDecl {
-                        name: "p".into(),
-                        ty: None,
-                        value: Expression::StructLiteral {
-                            name: "Point".into(),
-                            type_args: Vec::new(),
-                            fields: vec![(
-                                "x".into(),
-                                Expression::StringLiteral {
-                                    value: "bad".into(),
-                                    span: Span::dummy(),
-                                },
-                            )],
-                            span: Span::dummy(),
-                        },
-                        mutable: false,
-                        constant: false,
-                        span: Span::dummy(),
-                    }],
-                    expr: None,
-                    span: Span::dummy(),
-                },
-                public: false,
-                span: Span::dummy(),
-            },
-        ],
-        file_id: 0,
-    };
-
-    let errors = tc
-        .check_program(&program)
-        .expect_err("struct field type mismatch should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("field `x` for struct `Point` expects `i32`, found `str`")),
-        "expected field type diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn enum_variant_unknown_variant_is_error() {
-    let program = parse_program(
-        r#"
-Status: Ok, Err
-
-main = () void {
-    value = Status.Pending
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("unknown enum variant should fail");
-    assert!(
-        errors
-            .iter()
-            .any(|d| d.message.contains("enum `Status` has no variant `Pending`")),
-        "expected unknown variant diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn enum_variant_payload_type_mismatch_is_error() {
-    let program = parse_program(
-        r#"
-Maybe: Some(i32), None
-
-main = () void {
-    value = Maybe.Some("bad")
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("enum payload type mismatch should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("payload for enum variant `Maybe.Some` expects `i32`, found `str`")),
-        "expected payload type diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn assignment_to_immutable_binding_is_error() {
-    let program = parse_program(
-        r#"
-main = () void {
-    x = 1
-    x = 2
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("immutable assignment should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("cannot assign to immutable variable `x`")),
-        "expected immutable assignment diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn assignment_to_mutable_closure_parameter_is_allowed() {
-    let program = parse_program(
-        r#"
-main = () void {
-    mapper = (mut input: i32) i32 {
-        input = input + 1
-        input
-    }
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    tc.check_program(&program)
-        .expect("mutable closure parameter assignment should pass");
-}
-
-#[test]
-fn assignment_type_mismatch_is_error() {
-    let program = parse_program(
-        r#"
-main = () void {
-    x ::= 1
-    x = "bad"
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("assignment type mismatch should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("assignment to `x` expects `i32`, found `str`")),
-        "expected assignment type diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn invalid_field_access_is_error() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-main = () void {
-    p = Point { x: 1 }
-    y = p.y
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("invalid field access should fail");
-    assert!(
-        errors
-            .iter()
-            .any(|d| d.message.contains("type `Point` has no field `y`")),
-        "expected invalid field diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn implicit_integer_width_conversion_is_error() {
-    let program = parse_program(
-        r#"
-take_i64 = (value: i64) void {}
-
-main = () void {
-    x: i32 = 1
-    take_i64(x)
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("implicit integer conversion should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("argument 1 for `take_i64` expects `i64`, found `i32`")),
-        "expected integer conversion diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn implicit_float_width_conversion_is_error() {
-    use crate::ast::{Expression, Program};
-    let mut tc = TypeChecker::new();
-    let program = Program {
-        declarations: vec![Declaration::Function {
-            name: "take_f32".into(),
-            type_params: Vec::new(),
-            params: vec![ast::Param {
-                name: "value".into(),
-                ty: AstType::F32,
-                mutable: false,
-                span: Span::dummy(),
-            }],
-            return_type: Some(AstType::Void),
-            body: Expression::Block {
-                statements: Vec::new(),
-                expr: None,
-                span: Span::dummy(),
-            },
-            public: false,
-            span: Span::dummy(),
-        }],
-        file_id: 0,
-    };
-    tc.collect_declarations(&program.declarations);
-
-    let expected = tc.functions["take_f32"].params[0].1.clone();
-    assert!(!tc.types_compatible(&tc.resolve_type(&expected), &Type::F64));
-}
-
-#[test]
-fn unknown_root_std_module_call_is_error() {
-    let program = parse_program(
-        r#"
-{ io } = std
-
-main = () void {
-    io.nope("bad")
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("unknown std module function should fail");
-    assert!(
-        errors
-            .iter()
-            .any(|d| d.message.contains("undefined module function `io.nope`")),
-        "expected undefined module function diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn known_root_std_runtime_standins_remain_allowed() {
-    let program = parse_program(
-        r#"
-{ io } = std
-
-main = () void {
-    io.print("hello")
-    io.println("world")
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    tc.check_program(&program)
-        .expect("temporary root std io stand-ins should typecheck");
-}
-
-#[test]
-fn non_void_function_without_return_is_error() {
-    let program = parse_program(
-        r#"
-missing = () i32 {
-    x = 1
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("non-void fallthrough should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("function `missing` must return `i32` on all non-error paths")),
-        "expected missing return diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn enum_match_missing_variant_is_error() {
-    let program = parse_program(
-        r#"
-Color: Red, Green, Blue
-
-describe = (c: Color) StaticString {
-    c ?
-        | Red { "red" }
-        | Green { "green" }
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("non-exhaustive enum match should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("non-exhaustive match on `Color`: missing `Blue`")),
-        "expected non-exhaustive enum diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn enum_match_duplicate_variant_is_error() {
-    let program = parse_program(
-        r#"
-Color: Red, Green
-
-describe = (c: Color) StaticString {
-    c ?
-        | Red { "red" }
-        | Red { "again" }
-        | Green { "green" }
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("duplicate enum match arm should fail");
-    assert!(
-        errors
-            .iter()
-            .any(|d| d.message.contains("duplicate match arm for `Color.Red`")),
-        "expected duplicate enum arm diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn enum_match_unknown_variant_is_error() {
-    let program = parse_program(
-        r#"
-Color: Red, Green
-
-describe = (c: Color) StaticString {
-    c ?
-        | Red { "red" }
-        | Blue { "blue" }
-        | Green { "green" }
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("unknown enum match arm should fail");
-    assert!(
-        errors
-            .iter()
-            .any(|d| d.message.contains("enum `Color` has no variant `Blue`")),
-        "expected unknown enum arm diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn enum_match_payload_shape_is_checked() {
-    let program = parse_program(
-        r#"
-Maybe: Some(i32), None
-
-describe = (m: Maybe) StaticString {
-    m ?
-        | Some { "some" }
-        | None(value) { "none" }
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("enum match payload shape should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("match arm `Maybe.Some` requires a payload")),
-        "expected missing payload diagnostic, got {errors:?}"
-    );
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("match arm `Maybe.None` does not accept a payload")),
-        "expected forbidden payload diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn enum_match_wildcard_after_all_variants_is_redundant() {
-    let program = parse_program(
-        r#"
-Color: Red, Green
-
-describe = (c: Color) StaticString {
-    c ?
-        | Red { "red" }
-        | Green { "green" }
-        | _ { "fallback" }
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("redundant enum wildcard arm should fail");
-    assert!(
-        errors
-            .iter()
-            .any(|d| d.message.contains("redundant wildcard match arm")),
-        "expected redundant wildcard diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn enum_match_variant_after_wildcard_is_redundant() {
-    let program = parse_program(
-        r#"
-Color: Red, Green
-
-describe = (c: Color) StaticString {
-    c ?
-        | _ { "fallback" }
-        | Red { "red" }
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("enum variant after wildcard should fail");
-    assert!(
-        errors
-            .iter()
-            .any(|d| d.message.contains("redundant match arm for `Color.Red`")),
-        "expected redundant enum arm diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn bool_match_missing_arm_is_error_for_value_match() {
-    let program = parse_program(
-        r#"
-describe = (flag: bool) StaticString {
-    flag ?
-        | true { "yes" }
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("non-exhaustive boolean value match should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("non-exhaustive bool match: missing `false`")),
-        "expected non-exhaustive bool diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn bool_match_duplicate_arm_is_error() {
-    let program = parse_program(
-        r#"
-describe = (flag: bool) StaticString {
-    flag ?
-        | true { "yes" }
-        | true { "again" }
-        | false { "no" }
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("duplicate boolean match arm should fail");
-    assert!(
-        errors
-            .iter()
-            .any(|d| d.message.contains("duplicate match arm for `true`")),
-        "expected duplicate bool arm diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn match_arm_return_does_not_force_never_result_type() {
-    let program = parse_program(
-        r#"
-describe = (flag: bool) StaticString {
-    flag ?
-        | true { "early" }
-        | false { "late" }
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let typed = tc
-        .check_program(&program)
-        .expect("returning arm should not force match type to never");
-    let body = &typed.functions[0].body;
-    assert_eq!(body.ty, Type::Str);
-}
-
-#[test]
-fn types_compatible_basics() {
-    let tc = TypeChecker::new();
-    // Same types
-    assert!(tc.types_compatible(&Type::I32, &Type::I32));
-    // Numeric conversions require explicit casts except literal coercion.
-    assert!(!tc.types_compatible(&Type::I64, &Type::I32));
-    assert!(!tc.types_compatible(&Type::F32, &Type::F64));
-    // Unknown is permissive
-    assert!(tc.types_compatible(&Type::I32, &Type::Unknown));
-    // Named types are nominal and do not match unrelated concrete types.
-    assert!(tc.types_compatible(&Type::Named("UserId".into()), &Type::Named("UserId".into())));
-    assert!(!tc.types_compatible(
-        &Type::Named("UserId".into()),
-        &Type::Named("OrderId".into())
-    ));
-    assert!(!tc.types_compatible(&Type::Str, &Type::Named("StaticString".into())));
-    // Clear mismatch
-    assert!(!tc.types_compatible(&Type::I32, &Type::Str));
-    assert!(!tc.types_compatible(&Type::Bool, &Type::I32));
-}
-
-#[test]
-fn literal_coercion_in_var_decl() {
-    use crate::ast::{Expression, Program, Statement};
-    let mut tc = TypeChecker::new();
-    let program = Program {
-        declarations: vec![Declaration::Function {
-            name: "main".into(),
-            type_params: Vec::new(),
-            params: Vec::new(),
-            return_type: Some(AstType::Void),
-            body: Expression::Block {
-                statements: vec![Statement::VarDecl {
-                    name: "x".into(),
-                    ty: Some(AstType::I64),
-                    value: Expression::IntLiteral {
-                        value: 42,
-                        span: Span::dummy(),
-                    },
-                    mutable: false,
-                    constant: false,
-                    span: Span::dummy(),
-                }],
-                expr: None,
-                span: Span::dummy(),
-            },
-            public: false,
-            span: Span::dummy(),
-        }],
-        file_id: 0,
-    };
-    let result = tc.check_program(&program).unwrap();
-    // The variable should have type I64 (coerced from I32 literal)
-    let body = &result.functions[0].body;
-    match &body.statements[0].kind {
-        TypedStatementKind::VarDecl { ty, .. } => assert_eq!(*ty, Type::I64),
-        _ => panic!("expected VarDecl"),
-    }
-}
-
-#[test]
-fn resolve_string_type() {
-    let tc = TypeChecker::new();
-    // "String" as a named type should resolve to Type::String
-    assert_eq!(
-        tc.resolve_type(&AstType::Named("String".into())),
-        Type::String
-    );
-}
-
-#[test]
-fn resolve_slice_type() {
-    let tc = TypeChecker::new();
-    assert_eq!(
-        tc.resolve_type(&AstType::Slice(Box::new(AstType::I32))),
-        Type::Slice(Box::new(Type::I32))
-    );
-}
-
-#[test]
-fn infer_type_args_basic() {
-    let tc = TypeChecker::new();
-    // Generic function: identity<T>(x: T) -> T
-    let type_params = vec!["T".to_string()];
-    let params = vec![("x".to_string(), AstType::Named("T".into()))];
-    let arg_types = vec![Type::I32];
-    let subs = tc.infer_type_args(&type_params, &params, &arg_types);
-    assert_eq!(subs.get("T"), Some(&Type::I32));
-}
-
-#[test]
-fn substitute_type_basic() {
-    let tc = TypeChecker::new();
-    let mut subs = HashMap::new();
-    subs.insert("T".to_string(), Type::I32);
-    // T → I32
-    assert_eq!(
-        tc.substitute_type(&AstType::Named("T".into()), &subs),
-        Type::I32
-    );
-    // Ptr<T> → Ptr<I32>
-    assert_eq!(
-        tc.substitute_type(&AstType::Ptr(Box::new(AstType::Named("T".into()))), &subs),
-        Type::Ptr(Box::new(Type::I32))
-    );
-    // Non-generic type unchanged
-    assert_eq!(tc.substitute_type(&AstType::Bool, &subs), Type::Bool);
-}
-
-#[test]
-fn substitute_type_covers_all_composite_type_shapes() {
-    let tc = TypeChecker::new();
-    let mut subs = HashMap::new();
-    subs.insert("T".to_string(), Type::I32);
-
-    assert_eq!(
-        tc.substitute_type(
-            &AstType::RawPtr(Box::new(AstType::Named("T".into()))),
-            &subs
-        ),
-        Type::RawPtr(Box::new(Type::I32))
-    );
-    assert_eq!(
-        tc.substitute_type(
-            &AstType::MutPtr(Box::new(AstType::Named("T".into()))),
-            &subs
-        ),
-        Type::MutPtr(Box::new(Type::I32))
-    );
-    assert_eq!(
-        tc.substitute_type(&AstType::Slice(Box::new(AstType::Named("T".into()))), &subs),
-        Type::Slice(Box::new(Type::I32))
-    );
-    assert_eq!(
-        tc.substitute_type(
-            &AstType::Array {
-                elem: Box::new(AstType::Named("T".into())),
-                size: Some(3),
-            },
-            &subs,
-        ),
-        Type::Array {
-            elem: Box::new(Type::I32),
-            size: Some(3),
-        }
-    );
-    assert_eq!(
-        tc.substitute_type(
-            &AstType::Function {
-                params: vec![AstType::Named("T".into())],
-                ret: Box::new(AstType::Named("T".into())),
-            },
-            &subs,
-        ),
-        Type::Function {
-            params: vec![Type::I32],
-            ret: Box::new(Type::I32),
-        }
-    );
-}
-
-#[test]
-fn substitute_type_preserves_function_type_arguments_in_nested_generics() {
-    let mut tc = TypeChecker::new();
-    tc.structs.insert(
-        "Box".to_string(),
-        StructInfo {
-            name: "Box".to_string(),
-            fields: vec![("value".to_string(), AstType::Named("T".to_string()))],
-            field_defaults: HashMap::new(),
-            type_params: vec!["T".to_string()],
-            type_param_bounds: HashMap::new(),
-        },
-    );
-    let function_type = Type::Function {
-        params: vec![Type::I32],
-        ret: Box::new(Type::I32),
-    };
-    let mut subs = HashMap::new();
-    subs.insert("T".to_string(), function_type.clone());
-
-    assert_eq!(
-        tc.substitute_type(
-            &AstType::Generic {
-                name: "Box".to_string(),
-                type_args: vec![AstType::Named("T".to_string())],
-            },
-            &subs,
-        ),
-        Type::Struct {
-            name: "Box_fn_i32_ret_i32".to_string(),
-            fields: vec![("value".to_string(), function_type)],
-        }
-    );
-}
-
-mod generic_behaviors;
